backend/src/requirements.txt
mediapipe
opencv-python
fastapi
python-multipart
numpy
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
